fix: handle 32-bit platform build for uint comparison
On 32-bit platforms (ARM, x86), `uint` is 32 bits, so comparing
a `uint` value directly to `math.MaxInt64` causes a compile error
since the constant overflows. Widen `uint` to `uint64` before the
comparison.

Also add CI jobs for 32-bit cross-compilation (linux/arm, linux/386).

+++ filters/standard_filters.go
@@ -36,10 +36,8 @@ func isIntegerType(v any) bool {
 	case int, int8, int16, int32, int64, uint8, uint16, uint32:
 		return true
 	case uint:
-		// Check if uint value fits in int64 range
-		return val <= math.MaxInt64
+		return uint64(val) <= math.MaxInt64
 	case uint64:
-		// Check if uint64 value fits in int64 range
 		return val <= math.MaxInt64
 	default:
 		return false
